requirements.txt
Flask
gunicorn==20.1.0
heyoo
requests
google-cloud-speech